Mimic environment of original portals
The full path to where the application was cloned is stored in the
database by the portal. We are changing where the new ones will go
via the properties, but we need to still be able to access the existing
application and deployments -> create the necessary links there.

Also, for the first portal server we did, the install dir was different.
Create a link for that too.