Upgrade to the latest newrelic agent
The old version of newrelic agent and the portal code started being
incompatible it seems: we were seeing a `java.lang.ClassFormatError:
Duplicated LocalVariableTable attribute entry for 'e' in class file`
on startup of the portal api with the `-javaagent`, but it started
successfully without.

Switching to the new version of the newrelic agent fixes the issue.