Enable separation of python execution environments
Install virtualenv so the different applications running on the
portal can install what they need without interfering with the
other applications (they can already install anything, but it goes
in the global list so in theory there could be compatibility
issues).

The tool compatibility & availability issue in general will be
solved by moving the deployments to run on a docker image that the
users can specify, but the meantime this offers some isolation.